+++ package.json
@@ -15,7 +15,8 @@
     "prebuild": "npm run docs",
     "build": "ts-node scripts/index.ts && npm run css",
     "fix": "eslint 'scripts/**/*.ts' --fix",
-    "lint": "eslint 'scripts/**/*.ts'"
+    "lint": "eslint 'scripts/**/*.ts'",
+    "test": "jest"
   },
   "engines": {
     "node": ">=18.3.0"
@@ -32,6 +33,7 @@
     "@swc/helpers": "^0.4.14",
     "@types/dompurify": "^2.4.0",
     "@types/glob": "^8.0.0",
+    "@types/jest": "^29.2.5",
     "@types/jsdom": "^20.0.1",
     "@types/marked": "^4.0.8",
     "@types/node": "^18.11.17",
@@ -40,8 +42,11 @@
     "eslint": "^8.31.0",
     "eslint-import-resolver-typescript": "^3.5.2",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.2.0",
+    "jest": "^29.3.1",
     "regenerator-runtime": "^0.13.11",
     "sass": "^1.57.1",
+    "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.1.1",
     "typescript": "^4.9.4"
