+++ tsconfig.json
@@ -4,7 +4,7 @@
       "module": "ESNext",
       "allowJs": true,
       "strict": true,
-      "moduleResolution": "Node",
+      "moduleResolution": "NodeNext",
       "esModuleInterop": true,
       "baseUrl": "./",
       "skipLibCheck": true,
